Make adapter.notifyDataSetChanged() less aggressive.
At the moment, if we call notifyDataSetChanged() on our adapter, then
CardStackView unconditionally recreates all of the Views that are shown in
the list. In other words, there's never a case when adapter.getView() gets
called with a recycled contentView that can be reused for fresh content.

This patch lightens the operation of the onChanged() handler, where it
will now reuse any Views that are already created, and no longer recreate
the entire list.