Provide swiped position to onCardSwiped event
Whenever a card is swiped, this also captures the position
of the swiped card and emits it to the event. This allows
us to trust the position at the event handler, and not have
to inspect the cardLayoutManager directly. This is helpful
in the case where our event may trigger another runnable or
otherwise may not be able to trust that the layoutManager
has not already calculated another topPosition.